# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation TEA BLISS",
        "description": "API với Sanctum Authentication của Tea Bliss",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/{product_id}/avatar": {
            "get": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Lấy ảnh đại diện của sản phẩm",
                "operationId": "6d641dc91b89ad822e25d1ea1aa7c0bb",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID của sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trả về URL ảnh đại diện của sản phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy hình avatar thành công"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "https://yourdomain.com/storage/avatar.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy ảnh đại diện",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hình avatar không có"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{product_id}/images": {
            "get": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Lấy tất cả hình ảnh của sản phẩm",
                "operationId": "78529c3fd20dfb39622c990cbcbeb245",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID của sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách hình ảnh của sản phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy hình sản phẩm thành công"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://yourdomain.com/storage/image1.jpg"
                                                    },
                                                    "alt_text": {
                                                        "type": "string",
                                                        "example": "Hình ảnh sản phẩm"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy hình ảnh",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy hình sản phẩm thất bại"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}